feat(data_types): Copy over shares.rs from share-accounting-ext
It might be defined in SRI already but I'm not sure so copied it over. I'll be using it as a reference for traits I need to implement for a cashu keyset and other data primitives that are Cashu specific.

This Share struct is then part of a `message`. Messages sent on the wire and composed of `data_types`.